+++ src/components/chat-widget/chat-animations/diamond-shine/diamond-shine-keyframes.css
@@ -1,23 +1,27 @@
 .diamond-shine {
-  animation: pulsate 2s infinite;
+  animation: pulsate 1s infinite;
+  background-color: #ba0ef9a5;
+  /* Adjusted background color for more shine effect */
+  border-radius: 50%;
+  /* Assuming this is for a circular element */
 }
 
 @keyframes pulsate {
   0% {
-    box-shadow: 0 0 20px 0 rgba(174, 30, 227, 0.3), 0 0 30px 0 rgba(109, 75, 121, 0.2), 0 0 40px 0 rgba(109, 75, 121, 0.1);
-    opacity: 0.7;
+    box-shadow: 0 0 20px 5px #ba0ef9a5, 0 0 30px 10px rgba(65, 50, 71, 0.2), 0 0 10px 10px rgba(109, 75, 121, 0.1);
+    opacity: 0.9;
     /* Initial opacity */
   }
 
   50% {
-    box-shadow: 0 0 30px 0 rgba(174, 30, 227, 0.4), 0 0 40px 0 rgba(109, 75, 121, 0.3), 0 0 50px 0 rgba(109, 75, 121, 0.2);
-    opacity: 0.9;
+    box-shadow: 0 0 30px 10px #ba0ef9a5, 0 0 40px 15px rgba(65, 50, 71, 0.2), 0 0 20px 20px rgba(109, 75, 121, 0.3);
+    opacity: 1;
     /* Higher opacity at midpoint */
   }
 
   100% {
-    box-shadow: 0 0 20px 0 rgba(174, 30, 227, 0.3), 0 0 30px 0 rgba(109, 75, 121, 0.2), 0 0 40px 0 rgba(109, 75, 121, 0.1);
-    opacity: 0.7;
+    box-shadow: 0 0 20px 5px #ba0ef9a5, 0 0 30px 10px rgba(65, 50, 71, 0.2), 0 0 10px 10px rgba(109, 75, 121, 0.1);
+    opacity: 0.9;
     /* Return to initial opacity */
   }
 }
